Update index.md
Just fixed a typo on the last paragraph of the casting example, there was an "is" that should b an if
Also fixed a typo in the code example itself. On the second example, the one with dynamic typing, you use the variable player without declaring it, so I changed it to body, which is the parameter of the callback.